cleaning the repository
removing unused files,
adding a readme
adding a run file for the sgd matrix factorization
