--- utils.py
@@ -1,7 +1,0 @@
-import time
-
-def create_output():
-    pass
-
-def pprint(string):
-    pass
